requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
numpy==1.14.0
pandas==0.22.0
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
